PR3672: don't print "_vptr.foo" members
The only use I can think for the vtable pointer is to try to identify an
instance's subclass, but I don't think that's useful in the general
case.  We might want to think about how to expose that for power users
though, because our syntax doesn't allow "$p->_vptr.foo".

* tapsets.cxx (dwarf_pretty_print::recurse_struct_members): Filter out
  members which start with "_vptr.foo".