Don't resolve symbols unnecessarily and don't suppress things from backtrace.
We were always doing symbol lookups of addresses in a backtrace, even when
the user only wanted the addresses. If we couldn't find an symbol associated
with an address in the backtrace we just tried again and reported the next
address/symbol we could find. It really is better to always report all
addresses we think are part of a backtrace. It is more efficient, and more
truthful.

* runtime/sym.c (_stp_func_print): Don't report whether name lookup
  succeeded. Only lookup symbol names when requested.
* runtime/stack-x86_64.c (_stp_stack_print_fallback): Report all addresses
  found.
* runtime/stack-i386.c (_stp_stack_print_fallback): Likewise.
  (__stp_stack_print): Mark frame based addresses as inexact.