PR11785 (partial): add grace to tracepoint $$parms$
Tracepoint args with base types should not be dereferenced at all, but
the pretty-printing is not wise enough to know that.  Thankfully we were
already blocking the pretty-printing from such vars, but the error was
turning into an assertion failure in $$parms$.

So rather than asserting, for now just turn them into "=?" like dwarf
probes do with such failure.  Later we can try to make pretty printing
work with such base-typed tracepoint $vars.

* tapsets.cxx
  (tracepoint_var_expanding_visitor::visit_target_symbol_context):
  Don't assert that there aren't errors, just squash them.
* testsuite/semok/pretty2.stp: Exercise lots of tracepoints.