Sanitize before usage in unwinder if they come from task_pt_regs.
Introduce _stp_task_pt_regs_valid() that checks whether a full set
of user registers was saved. Currently assumes yes for i386 and no otherwise.
If not all registers were saved create a copy for the unwinder with
everything zeroed out except the (arch specific, x86_64 only for now)
known to be valid registers (would be nicer to do this as initial
unwinder state, so we could mark them undefined, but the current
form at least prevents leaking any data that might still be saved
in the pt_regs that task_pt_regs() points to).

* runtime/compatdefs.h (_stp_task_pt_regs_valid): New function.
* runtime/stack.c (_stp_stack_print): Take and pass uregs_valid argument.
  (_stp_stack_sprint): Likewise.
* runtime/stack-arm.c (__stp_stack_print): Likewise.
* runtime/stack-i386.c (__stp_stack_print): Likewise.
* runtime/stack-ia64.c (__stp_stack_print): Likewise.
* runtime/stack-ppc.c (__stp_stack_print): Likewise.
* runtime/stack-s390.c (__stp_stack_print): Likewise.
* runtime/stack-x86_64.c (__stp_stack_print): Likewise.
* tapset/context-unwind.stp (print_backtrace): Pass zero for valid.
  (sprint_backtrace): Likewise.
  (backtrace): Likewise.
* tapset/ucontext-unwind.stp (print_ubacktrace): Check and pass uregs_valid.
  (sprint_ubacktrace): Likewise.
  (print_ubacktrace_brief): Likewise.
  (ubacktrace): Likewise.
* runtime/unwind/i386.h (arch_unw_init_frame_info): Accept sanitize argument.
* runtime/stack-x86_64.c (arch_unw_init_frame_info): Likewise and actually
  sanitize the pt_regs copy.