Add all the weirdness to compat-%p
I thought the old %p behavior was as simple as %#.8x or %#.16x, but it
was even weirder than that, as the ptr.exp test reveals.  The compat
path is now able to pass that test, and a new test makes sure of the new
code behaves as expected too.

* runtime/vsprintf.c (_stp_vsnprintf): compat-%p is better preserved.
* testsuite/systemtap.printf/ptr.exp -> ptr-1.2.exp: Keep as a compat
  test, and add a check that %.Np precision is ignored.
* testsuite/systemtap.printf/ptr-1.3.exp: New, make sure that %p really
  does act like %#x now, with various field_width.precision combos.