Anchor bz10078.exp's regexp patterns
If we are unlucky in buffering, then the test output could be read
into expect as a single buffer.  Then one of our regexps could match the
end of the buffer and consume the whole thing, even if we also wanted to
match something earlier in the buffer.

The solution is to anchor the patterns to the beginning of the buffer
with '^', so we're sure to match everything we can.