Make sure Dwarf_Ops don't have garbage
We synthesize a Dwarf_Op in a couple of places, but we were only setting
the fields we care about, ".atom" and ".number".  The garbage in
".number2" would survive into "DWARF expression:" comment in loc2c,
which made it look unique and subverted caching.

Now we memset-0 the Dwarf_Op before proceeding.