PR11911: Include in -L vars which are partially available
Mostly this is for aggregate types, like a struct type where we can't
translate $foo, but $foo->bar works just fine.  So instead of seeing if
translation works without error, we now look for location info
ourselves.  We list all $vars which are present at this address, even
if not "fully" available.

This also adds verbosity on errors, for PR11910.

* tapsets.cxx (dwarf_derived_probe::saveargs): Check for locations
  manually instead of trying translation.
* testsuite/semok/bz11911.stp: New test adapted from Roland.