Resolve statement scopes lexically instead of by raw PC
When we have a statement PC within a function that we wish to probe,
it's not reliable to use getscopes(PC) for variable scoping.  This is
because there may also be inline instances at the PC, and so you can end
up in a different scope altogether.

So now we can traverse the DIEs within a function looking for the
innermost containing the PC, avoiding those nested DIEs which are
actually inlines.  This also lets us get rid of the hack that made
function prologues do a PC search, so we should have more reliable
scoping all around.

* dwflpp.cxx (dwflpp::inner_die_containing_pc): New, scan lexical scopes
  for the innermost DIE that contains a PC, avoiding inlines.
  (dwflpp::iterate_over_labels): Use the above to find the lexical scope
  containing the label's address.
* tapsets.cxx (query_srcfile_line): Use the above to find the closest
  DIE for each statement probe.
  (dwarf_var_expanding_visitor::getscopes): Always call getscopes on the
  given scope_die, so we don't have the trouble that a PC search entails.