List variables from containing scopes too
Now we iterate through the lexical scopes up to but not including the
file level, so we can find variables in containing scopes too.  This is
most significant for statement probes to list everything available.
Excluding file-level means that globals aren't listed, but we could
consider a more verbose version to include those too.

* tapsets.cxx (dwarf_derived_probe::saveargs): Iterate through scopes.