PR12121: distinguish struct/typedef homonyms
For example:
  typedef struct foo foo;
  foo x;

This kind of "repeated" typedef used to cause bad things for us if the
struct definition is not in the current CU.  When resolving $x->bar, we
start at the typedef-foo and dereference to the struct-foo.  That's just
a declaration, so we use declaration_resolve to look for the struct
definition.  But since we were looking by the name "foo" only, we would
find the typedef again, getting us nowhere.

The fixed code does declaration_resolve on "struct foo" in the global cache,
so in this case we won't find anything locally, and thus we continue looking
in other CUs.  If the definition isn't found elsewhere, we're stuck anyway.

* dwflpp.cxx (dwflpp::global_alias_caching_callback): Cache type names
  with a relevant prefix: struct, union, etc.  Classes are normalized to
  be "struct" too, as that's a very blurry line.
  (dwflpp::declaration_resolve(die)): New variant to resolve a specific
  die -- used so we know what exact prefix we should look for.
* parse.cxx (parser::parse_target_symbol): Don't strip the prefix from
  @cast types anymore, so the user can be precise about what they want.
* tapsets.cxx (dwarf_cast_query::handle_query_module): Normalize "class"
  to "struct".  If the user provides a bare name that fails to resolve,
  then try it again with the various prefixes.
* testsuite/systemtap.base/unresolved-struct-typedef.*: New test.