uprobes: Fix post_ssout handling of "rep ret" on x86
That odd sequence is apparently used to coerce better behavior from the
branch predictor on AMD K8.  GCC does this, so we need to be prepared to
deal with it.

In uprobe_post_ssout, most instructions just need a relative %ip fixup
after single-stepping out-of-line.  A few are special though, either
because their new %ip is not relative, or because their return address
on the stack needs that relative SSOL adjustment, or even both.  This
oddball "rep ret" is such a case, but we were missing the proper fixup
due to the unexpected prefix.

This patch moves that ip-fixup decoding to the validation phase in
setup_uprobe_post_ssout, saving the flags UPFIX_RETURN and UPFIX_ABS_IP
as needed.  In this same step, we reject probing these instructions if
they have a prefix, as we haven't determined the correct behavior for
them.  For "rep ret" we make an explicit exception, since we know that
behaves just like a bare "ret".