Audit some minus/dash/hyphen use in manpages
Groff treats "-" as a hyphen, U+2010.  When the intention is to produce
a literal ASCII minus sign, U+002D, it must be "\-".  This matters for
command-line options and examples to be copyable from the manpage.

I've updated cases that should obviously be escaped.  At a glance, I
think we have other cases that could be normal hyphens, not escaped, but
that's less of an issue.