Disable systemtap.examples/process/noptrace.stp on RHEL4 systems (PR12498)
The debug information on RHEL4 kernels appears to be wrong. This causes
the noptrace.stp example to write into the wrong location. Rather than the
ptrace syscall exiting immediately, the ptrace syscall continues and the
test hangs on RHEL4 systems.  The noptrace.stp test is disabled on RHEL4
systems to work around this issue.