stapsh: A systemtap "shell" to aid remote execution
This stapsh agent is a wrapper meant to be automatically invoked through
an ssh connection to transfer systemtap modules to remotes hosts and
invoke staprun on them.  The primary motivation of stapsh is make sure
all the exit cases can happen cleanly, especially with sending signals
to the remote and cleaning up the processes and temporary files.

This is meant to replace the code based on multiple connections with an
ssh ControlMaster.  For now, that code has only been #if'ed out in
case we still want it for talking with "legacy" remotes.

* runtime/staprun/stapsh.c: New, a basic command shell for uploading and
  remotely launching systemtap scripts.
* runtime/staprun/Makefile.am: Add stapsh.
* runtime/staprun/Makefile.in: Regen.
* systemtap.spec: Package stapsh in systemtap-runtime.
* remote.cxx (class stapsh): New, handle communication with a stapsh
  remote via any given file descriptors.
  (class direct_stapsh): New, launch a direct stapsh instance to run
  through.  This way we can add stapsh to the testsuite without
  requiring any remote access configuration.
  (class ssh_remote): Block out the old implementation, and rewrite
  as an inheritor of stapsh.
  (remote::create): Create direct_stapsh for "--remote stapsh".
  (remote::run): Add a polling loop during runtime, where each remote
  can add fds to watch and then respond.
* util.cxx (stap_spawn_piped): Allow piped stdin too.
  (stap_system_read): Update call to stap_spawn_piped.