PR12591: Setup the signal handler earlier
We were calling setup_signals() in passes_0_4(), but --use-server
diverts that execution path, so signaled cleanup wasn't working in that
case at all.  It's also possible with multiple --remotes that
passes_0_4() will be called multiple times, though that shouldn't really
hurt signal handling.  This patch moves setup_signals() to main(), so it
is called once early on.

* main.cxx (passes_0_4): Don't call setup_signals() here.
  (main): Call setup_signals() here instead.