Add a testcase for stapsh remotes
We can't generally test true remotes in the testsuite without previously
setting up ssh access on a machine.  But we can still check that the
stapsh mechanism works in general.  This runs with --remote=stapsh, and
checks that stapsh exists in the process hierarchy and a signal to stap
only is propagated through stapsh.

* testsuite/systemtap.base/stapsh.*: New.