nsscommon: Heed what fgets returns
My F14 gcc rightfully warned that a return value of fgets was unused.
It may return NULL on error or EOF, in which case we should free our
memory and return NULL to NSS as well.  I also tweaked the newline stomp
to tolerate missing newline, e.g. with oversized input, although that
case it not likely going to work anyway.

* nsscommon.c (nssPasswordCallback): Check fgets' return for success.