remote: Try the legacy ssh code if stapsh isn't found
The ssh remote command is restructured so that if stapsh isn't found in
the remote PATH, it will exit 127 immediately.  On the local side, when
we fail to read the stapsh uname response, we can check the exit status
to know stapsh wasn't found.  Then we can try again with the old ssh
code with a ControlMaster.  However, I'm limiting the fallback to only
when there's a single remote, as signals and cleanup aren't really
handled as well.  (That's the point of stapsh...)

* remote.cxx (ssh_remote::create): New static function which tries a
  self-instantiation first, then if that fails to find stapsh, it tries
  again using ssh_legacy_remote instead.
  (ssh_remote::connect): Renamed from remote.  Now using a "type -P"
  trick to silently determine if stapsh is present, by exit code alone.
  (ssh_legacy_remote): Un-#ifdef'ed the old ssh_remote code, to get new
  life as our legacy access.
  (remote::create): Get ssh through ssh_remote::create.