Add a workaround for old systems without ppoll
A coarse approximation of ppoll is added using sigprocmask and poll, as
suggested in the ppoll manpage, but it doesn't carry the same atomic
guarantee.  I've forced it to have a non-infinite polling timeout, so we
won't block forever.

NB: This only matters on the client connecting to a remote ssh host.

* configure.ac: Check for the function ppoll.
* configure, config.in: Regenerate.
* remote.cxx (ppoll): Fill in for missing ppoll, with a healthy comment
  explaining why I think this implementation is Good Enough™.