module shutdown: be verbose upon a hang
In the generated module_exit() function, there exists an infinite loop
that waits for all probe handlers that might still be struggling along
to shut down, before authorizing removal of the probe module.  If
something is stuck though, it will stay stuck without any diagnostics
short of a hung stapio/staprun process that's sitting in the kernel.

This patch causes a message to be printk KERN_ERR'd, in case the
shutdown synchronization takes more than a second.

* translate.cxx (emit_module_exit): Print a message if holdon-spinning
  for more than a second.