Add code to deal with log file rotation.
Depending on how the log file is rotated, it will often end up backed
by a new inode (new log file) as the previous one is renamed (usually
with yesterdays log timestamp).  When this happens, we'll stop getting
events ... unfortunate.

So, make use of the stat data queried at start of fetch now to ensure
we detect this scenario, close the old file and switch to the new one
seamlessly.  In the process, we need to make sure we handle the case
where the file doesn't exist.