uprobes: impedance match insn tables with test_bit()
The kernel's test_bit expects its bitmap to be const volatile, but we
had ours as simply const.  On Fedora 15 with gcc 4.6, compiling uprobes
gave a few warnings like this:

    arch/x86/include/asm/bitops.h:319:2: warning: use of memory input
    without lvalue in asm operand 1 is deprecated [enabled by default]

That line is the asm statement in variable_test_bit().

The symptom noticed was that handle_riprel_insn was reading need_modrm:0
for opcode 0x89, when our table says it should be 1.  Who knows what
other havok ensued...

When our instruction tables are set const volatile to match test_bit(),
the warning goes away, and need_modrm is now computed correctly.