remote: Disambiguate the private target names
It's conceivable, however unlikely, that a user may have an actual host
named "direct" or "stapsh", which would conflict with our internal
methods if used as a --remote.  Such a user could say "ssh://direct" to
be explicit, but we can also hide ours a little better.  Those internal
names are now tested as proper URI schemes, e.g. "direct:...", so they
should never conflict with a user's legitimate target.

* remote.cxx (remote::create): Test for "direct" and "stapsh" only as
  the scheme of a decoded URI.
* main.cxx (main): Use "direct:" for non-remote use.
* testsuite/systemtap.base/stapsh.exp: Use "stapsh:" for testing.