Add initial support for inode-based uprobes
This adds support for placing regular userspace probes using the new
inode+offset API being developed for the upstream kernel.  This includes
probing functions, statements, and SDT markers, but return probes aren't
yet supported in the new API.  A lot of the finer details of systemtap's
userspace runtime still needs work too, but this is a functional start.

* runtime/uprobes-inode.c: New, basic registration code to lookup
  filename inodes and connect uprobes using the new API.
* tapsets.cxx (kernel_supports_inode_uprobes): New, guess whether this
  is an inode-uprobes kernel based on CONFIG values.
  (dwarf_builder::build): Disallow userspace return probes.
  (uprobe_derived_probe::join_group): Only trigger task_finder and the
  manual uprobes model for the old style of uprobes.
  (uprobe_builder::build): Disallow absolute-address userspace probes.
  (uprobe_derived_probe_group::emit*): Split into inode/utrace variants.