Properly report the return code (rc) for registered hw breakpoint (PR12947)
In the case of an error, the return code (rc) was not set when
registering a hw breakpoint. This lead to the the SystemTap instrumentation
module trying to unregister a non-existant hw breakpoint and causing a
kernel oops.  This patch ensure that the return code (rc) is set
and errors in hw breakpoint registration are handled correctly.