testsuite framework: stap_run2 quoting fix
When invoking stap_run2 with multiple optional stap arguments, the
extra arguments are captured in the tcl list $args.  When stap_run2
calls down to stap_run3 to do the real work, it gets packaged as a
single quoted string instead of the original list of options.
We need to unpack this list to pass it on, e.g. via tcl eval.

This impacts test cases that pass multiple parameters, such as
memory1.exp, const_value.exp, process_by_cmd.exp.