CVE-2011-2503: read instead of mmap to load modules
As staprun is preparing to load a kernel module, we first mmap the whole
module as MAP_PRIVATE. Then we proceed with our security checks,
including a trusted-signature validation on the mapped region, and if
all checks out, we'll call init_module() with that same mapped region.

However, MMAP(2) says of MAP_PRIVATE, "It is unspecified whether changes
made to the file after the mmap() call are visible in the mapped
region."  From my testing, it appears that file changes do indeed show
up in our mapped memory.  This means we have a TOCTOU race between
verifying the signature of that memory and then calling init_module().

By using read() instead of mmap(), we ensure that we have a fully
private copy of the module to verify and load, without fear of change.