PR12895: Use NOSTDINC_FLAGS in kernel stapconf checks
We should never be looking in /usr/include/ for headers when building
for the kernel.  This particularly bit us in a case where RHEL6 gained
blk_types.h in newer kernels.  So if the system had kernel-headers.rpm
with this new header in /usr/include/, but was still running an older
kernel that lacked it in /lib/modules/`uname -r`/build/, then we'd
misidentify that header's availability in stapconf.

* buildrun.cxx (compile_pass): Add NOSTDINC_FLAGS to CHECK_BUILD.