Guard regparm in probe context with STAP_NEED_REGPARM.
regparm is only used on i386 (and x86_64 when probing 32bit) for
signaling what the parameter packing is for the probed function
as set in register.stp. So only include it in the probe context
struct when actually used in a stap script.

Also updated docs and fixed signature on other arches.