Add probe_type to common probe context.
This exposes the type of probe handler for use in runtime functions,
so they can better anticipate which common probe context is valid at
runtime. Also introduces a new context tapset function probe_type()
that exposes a (informational) probe handler string to scripts. Mainly
useful for debugging, when the user wants to find out which kind of
low level probe handler a high-level probe translates to for a particular
kernel/systemtap version.