PR11209: Add cpu_clock and local_clock
This adds cpu_clock_ns(cpu), local_clock_ns(), and their coarser
variants.  These are monotonic wallclock time sources, rather than the
absolute time-since-epoch sort.  On kernels too old for these functions,
it falls back to stap's gettimeofday.

A kernel supports this only if it has the functions exported, and if it
is deemed new enough to have the patches that make it NMI-safe.  See the
comment in timestamp_monotonic.stp for details.

* buildrun.cxx: Add export checks for cpu_clock and local_clock.
* tapset/timestamp_monotonic.stp: Define the new functions.
* testsuite/buildok/timestamp_monotonic-embedded.stp: Build them.
* doc/SystemTap_Tapset_Reference/tapsets.tmpl: Include the new docs.