Fix the uprobes.ko path when used with remotes
When uprobes.ko is sent through stapsh, it will go in a temp path that
only stapsh knows.  However, we were sending -u/path/to/uprobes.ko with
a client side path, which would only work for --remote localhost.  :/
Now the path is sanitized to ./ for the stapsh working directory.

* buildrun.cxx (make_run_command): Take a remotedir instead of a module
  name.  When set, use it to normalize both the uprobes and module path.
  Also check versions that a uprobes path is even supported (>=1.4).
* remote.cxx (stapsh::start): The remotedir is simply "." for stapsh.
  (ssh_legacy_remote:start): The remotedir is the known from the prior
  mktemp -d.  Also start sending the module signature across, add a note
  why we unfortunately can't send uprobes in legacy mode, and specify
  the staprun command without any path in case the remote is different.