PR6961 - backtrace from non-pt_regs probe context
When printing a kernel backtrace use the kernel dump_trace if available.
If the kernel is compiled with frame pointers we can guess the stack
pionter pretty reliably, otherwise let dump_stack figure it out and
skip some of the framework stack frames before starting the print the
remaining stack.