PR3217: Merge similar compiled-printfs
It's pretty common for tapset wildcards to expand to lots of probes with
slightly different handlers (e.g. syscall.*), and we'd like for the
newly expanded printf code to not be duplicated so much.

This patch keeps a hash in the translator of visited format strings, so
we can output a compiled version of that only once.  This requires some
new argument space in the context, but we can optimize by realizing that
(1) printfs never nest when arguments are precomputed, (2) string
arguments are immutable and need not be copied, and (3) return strings
for sprintf can write directly to the caller's tmpvar.

* translate.cxx (c_unparser::compiled_printfs): New map for storing
  pair(print_to_stream, format_string) -> compiled printf name.
  (c_unparser::emit_compiled_printfs): New, emit the expanded function
  body for every compiled printf.
  (c_unparser::emit_compiled_printf_locals): New, emit a union-struct of
  context locals for all compiled printfs.
  (c_unparser::emit_common_header): Call emit_compiled_printf*.
  (c_unparser::declare_compiled_printf): New, called to declare that we
  need this printf -- add it to the map if it's not already there.
  (c_tmpcounter::visit_print_format): Declare printfs as we see them.
  (c_unparser::get_compiled_printf): New, get the name for a previously
  declared printf function.
  (c_unparser::visit_print_format): Get the compiled printf function.