PR13272: Fix quoted strings in the printf shortcut path
For printf strings with no arguments, we usually skip all the printf
path altogether and just use _stp_print.  However, the compiled-printf
changes affected the level of string escaping, which broke for a case
like printf("\"quote\\this\"\n").  We must make sure that this shortcut
path uses an output-quoted string literal.

* translate.cxx (c_unparser::visit_print_format): Generate an output-
  quoted format_string too, for use by the _stp_print shortcut.
  (preprocess_print_format): The "%s\n" case shouldn't alter components.
* testsuite/buildok/printf.stp: Add "%s" and "%s\n" shortcut tests too.