tracepoints: Skip identical headers from the build and source trees
In preparing the tracequery source, we normalize paths containing
"include/" to just the suffix part, so the header will be found by the
normal -I path.  However, the debuginfo source tree usually contains
at least the same headers found in the build tree.  We were thus
building effectively the same tracequery twice for headers that are
present in both the build and source trees.  Stop that!

* tapsets.cxx (tracepoint_builder::init_dw): Build the header list with
  knowledge of the "include/" normalization to avoid extra work.