Tweak cache logic to group removals by hash value
We used to glob the cache for *.ko, *.so, *.h, and *.log, implicitly
assuming that this was all that was cached.  For the .ko case we also
grouped .c and .sgn to be removed all together.  But when any new
extension came along, the cache cleanup ignored it.

Now we glob instead for any file in the cache that looks like is has a
hashed name, and we group removals for all files that appear to have the
same hash value, regardless of prefix or suffix.  This will hopefully
make us more robust against future file-caching changes.

* cache.cxx (cache_ent_info): Becomes a vector of paths, with no
  knowledge needed for modules.  Use better types for size and mtime.
  (clean_cache): Use a new glob that looks for hash-like files, with 32
  hex digits in a row.  Use a regex to find common HASH_LEN substrings
  in all those filenames.  Build the cache_ent_info groups by HASH_LEN,
  and proceed to delete old entries until below quota.